Fix packaging to include the type re-exports.

+++ script/build/lib/clean-types.ts
@@ -1,3 +1,4 @@
+import { existsSync } from "node:fs";
 import { rm, rmdir } from "node:fs/promises";
 import { join } from "node:path";
 import { packageNames } from "../common/package-info";
@@ -6,6 +7,8 @@ const TYPESCRIPT_DECLARATION_INDEX = "index.d.ts";
 
 for (const packageName of packageNames) {
   const indexFileName = join(packageName, TYPESCRIPT_DECLARATION_INDEX);
-  await rm(indexFileName);
-  await rmdir(packageName);
+  await rm(indexFileName, { force: true });
+  if (existsSync(packageName)) {
+    await rmdir(packageName);
+  }
 }
